+++ java/tud/gamecontroller/players/XXXXPlayer/Model.java
@@ -26,7 +26,10 @@ import tud.gamecontroller.game.StateInterface;
 import tud.gamecontroller.game.impl.JointMove;
 import tud.gamecontroller.term.TermInterface;
 
-import java.util.*;
+import java.util.ArrayDeque;
+import java.util.Collection;
+import java.util.HashSet;
+import java.util.Stack;
 
 /**
  * Holds a model of how the true state of the game may look given the percepts seen so far
@@ -69,7 +72,7 @@ public class Model<TermType extends TermInterface> implements Cloneable{
     public Stack<Collection<TermType>> getPerceptPath() { return this.perceptPath; }
     public ArrayDeque<Integer> getActionPathHashPath() { return this.actionPathHashPath; }
 //    public int getActionPathHash() { return this.actionPathHash; }
-public int getActionPathHash() { return this.actionPathHashPath.peekLast() == null ? -1 : this.actionPathHashPath.peekLast(); }
+    public int getActionPathHash() { return this.actionPathHashPath.peekLast() == null ? -1 : this.actionPathHashPath.peekLast(); }
 //    public int getPreviousActionPathHash() { return this.previousActionPathHash; }
     public int getPreviousActionPathHash() {
         Integer last = this.actionPathHashPath.pollLast();
@@ -82,10 +85,10 @@ public int getActionPathHash() { return this.actionPathHashPath.peekLast() == nu
         return this.statePath.isEmpty() ? match.getGame().getInitialState() : this.statePath.peek();
     }
     public Collection<TermType> getLatestExpectedPercepts() { return this.perceptPath.peek(); }
-    public int getNumberOfPossibleActions() {
-        int total = 1;
+    public double getNumberOfPossibleActions() {
+        double total = 1.0;
         for(int num : this.numberOfPossibleActionsPath) {
-            total *= num;
+            total *= (double) num;
         }
         return total;
     }
